+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Option to download all annotations.
+- Option to read annotations from a local file instead of online.
 
 ### Fixed